+++ src/app/providers/router/index.tsx
@@ -3,10 +3,6 @@ import { BaseLayout } from '@/app/layouts/base-layout';
 import App from '@/App';
 import { DepositCurvePage } from '@/pages/DepositCurvePage/ui/DepositCurvePage';
 
-// Импорты ваших страниц (создайте их по необходимости)
-// import { AboutPage } from '@/pages/about';
-// import { ContactPage } from '@/pages/contact';
-
 const router = createBrowserRouter([
   {
     path: '/',
@@ -14,19 +10,15 @@ const router = createBrowserRouter([
     children: [
       {
         index: true,
-        element: <App />, // Ваш главный компонент
+        element: <App />,
       },
       {
         path: 'curve',
         element: <DepositCurvePage />,
       },
-      // {
-      //   path: 'contact',
-      //   element: <ContactPage />,
-      // },
       {
         path: '*',
-        element: <div>404 - Страница не найдена</div>, // Замените на ваш компонент 404
+        element: <div>404 - Страница не найдена</div>,
       },
     ],
   },
