+++ src/features/deposit-curve-chat/model/types.ts
