+++ src/features/view-trade-ui/ui/ScreenshotPreview.tsx
