+++ src/lib/formatters/formatDate.ts
@@ -1,0 +1,9 @@
+
+export function formatDate(iso: string) {
+  try {
+    const d = new Date(iso);
+    return d.toLocaleString();
+  } catch {
+    return iso;
+  }
+}
