+++ src/pages/HomePage/index.tsx
@@ -1,6 +1,6 @@
 import SettingsPanel from "@/features/settings-panel/ui/SettingsPanel";
 import StatsPanel from "@/widgets/stats-panel/ui/StatsPanel";
-import ScreenshotPreview from "@/widgets/ScreenshotPreview";
+import ScreenshotPreview from "@/features/view-trade-ui/ui/ScreenshotPreview";
 import TradesList from "@/widgets/trade-list/ui/TradeList";
 
 export default function HomePage() {
