+++ src/shared/hooks/use-router.ts
